Use localstack endpoint in localdev
